+++ changelog.md
@@ -4,7 +4,12 @@ This project does NOT (yet) adhere to [Semantic Versioning](http://semver.org/).
 It is running as an Alfa/Beta, once v1 is reached we will start supporting semver.
 For now, please use all minor versions as major versions and each build version will be a minor version.
 
-## [0.4]
+
+## [Unreleased]
+### Added
+- `GetOrderScheduleOptionsRequest` & `OrderService::getScheduleOptions()`
+
+## [0.4.0]
 ### Removed
 - `amount` from `CreateCreditCardPaymentRequest` & `CreateCashOnDeliveryPaymentRequest`
 
@@ -22,7 +27,7 @@ For now, please use all minor versions as major versions and each build version 
 - `GetPaymentRequest` & `PaymentService::getPayment()`
 - `OrderPaymentStatus`
 - `contactNumber` to `OrderItemMerchant`
-- GetProductTag request and handler
+- `GetProductTagRequest` & `ProductService::getProductTag()`
 
 ### Fixed
 - Fixed "scheduledTime" body parameter name on create order request